fixed all issues

+++ src/app/mdules/blogs/blog.controllers.ts
@@ -3,13 +3,25 @@ import sendResponse from '../../utils/sendResponse';
 import catchAsync from '../../utils/catchAsync';
 import httpStatus from 'http-status';
 import AppError from '../../errors/AppError';
+import { User } from '../user/user.model';
+import { Types } from 'mongoose';
 
 // Create a new blog
 export const createBlog = catchAsync( async (req, res) => {
+   
+    const { email, role} = req.user;
+    const author = await User.findOne({ email, role} , { id: 1 })
+
+
+    if(!author){
+      throw new AppError(httpStatus.BAD_REQUEST, "Unauthorized access!")
+    }
 
     const blogData = req.body; // Get blog data from the request body
-    const result = await BlogServices.createBlog(blogData); // Call service to create blog
 
+    blogData.author = author._id
+    const result = await BlogServices.createBlog(blogData); 
+ 
     sendResponse(res, {
       statusCode: httpStatus.CREATED,
       success: true,
@@ -22,8 +34,7 @@ export const createBlog = catchAsync( async (req, res) => {
 // Get all blogs with optional filters
 export const getBlogs = catchAsync( async (req, res) => {
    
-   console.log( "user: ",req.user)
-    const result = await BlogServices.getAllBlogsFromDB(); // Call service to get blogs
+    const result = await BlogServices.getAllBlogsFromDB(req.query); 
 
     sendResponse(res, {
       statusCode: httpStatus.OK,
@@ -54,16 +65,24 @@ export const updateBlogById = catchAsync( async (req, res) => {
 
     const { id } = req.params; // Get blog ID from URL params
     const updateData = req.body; // Get updated data from request body
-    
+    const { email, role } = req.user;
+    const author = await User.findOne({ email, role});
+
+    if(!author){
+      throw new AppError(httpStatus.BAD_REQUEST, "Unauthorized access!")
+    }
+
     const isExistBlog = await BlogServices.getBlogById(id); 
     if(!isExistBlog){
       throw new AppError(httpStatus.NOT_FOUND, "The blog is not found!")
     }
-    console.log("user: ", req.user)
+
+    if(role !== 'admin' && !(author._id as Types.ObjectId).equals(isExistBlog.author._id)){
+      throw new AppError(httpStatus.BAD_REQUEST, "Unauthorized access!")
+    }
 
     const result = await BlogServices.updateBlogById(id, updateData); // Call service to update blog
 
-    console.log("result: ", result)
     sendResponse(res, {
       statusCode: httpStatus.OK,
       success: true,
@@ -78,7 +97,46 @@ export const updateBlogById = catchAsync( async (req, res) => {
 export const deleteBlogById = catchAsync( async (req, res) => {
 
     const { id } = req.params; // Get blog ID from URL params
+    const { email, role } = req.user;
+    const author = await User.findOne({ email, role});
 
+    if(!author){
+      throw new AppError(httpStatus.BAD_REQUEST, "Unauthorized access!")
+    }
+
+    const isExistBlog = await BlogServices.getBlogById(id); 
+    if(!isExistBlog){
+      throw new AppError(httpStatus.NOT_FOUND, "The blog is not found!")
+    }
+
+    if(role !== 'admin' && !(author._id as Types.ObjectId).equals(isExistBlog.author._id)){
+      throw new AppError(httpStatus.BAD_REQUEST, "Unauthorized access!")
+    }
+
+    
+    // Delete the blog from the database
+    const result = await BlogServices.deleteBlogById(id);
+
+    sendResponse(res, {
+      statusCode: httpStatus.OK,
+      success: true,
+      message: 'Blog deleted successfully',
+      data: result,
+    });
+
+});
+
+
+// admin can delete any blog
+export const adminDeleteBlogById = catchAsync( async (req, res) => {
+
+    const { id } = req.params; // Get blog ID from URL params
+
+    const isExistBlog = await BlogServices.getBlogById(id); 
+    if(!isExistBlog){
+      throw new AppError(httpStatus.NOT_FOUND, "The blog is not found!")
+    }
+    
     // Delete the blog from the database
     const result = await BlogServices.deleteBlogById(id);
